FEAT:  Firebase Config 빈 생성 에러 해결
