--- src/main/java/com/coding/yo/config/FirebaseConfig.java
@@ -1,26 +1,0 @@
-package com.coding.yo.config;
-
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseOptions;
-import com.google.firebase.auth.FirebaseAuth;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.io.IOException;
-
-@Configuration
-public class FirebaseConfig {
-
-
-    @Bean
-    public FirebaseAuth firebaseAuth() throws IOException {
-        FirebaseOptions options = FirebaseOptions.builder()
-                .setCredentials(GoogleCredentials.getApplicationDefault())
-                .build();
-        FirebaseApp.initializeApp(options);
-        return FirebaseAuth.getInstance(FirebaseApp.getInstance());
-
-
-    }
-}
