+++ src/main/java/com/coding/yo/security/config/FirebaseConfig.java
@@ -1,0 +1,42 @@
+package com.coding.yo.security.config;
+
+import com.google.auth.oauth2.GoogleCredentials;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.FirebaseOptions;
+import com.google.firebase.auth.FirebaseAuth;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.ClassPathResource;
+
+import java.io.FileInputStream;
+import java.io.IOException;
+
+@Configuration
+public class FirebaseConfig {
+//    @Bean
+//    public void firebaseAuth() throws IOException {
+//        FileInputStream key = new FileInputStream("src/main/resources/codingyo-firebase-adminsdk.json");
+//        FirebaseOptions options = FirebaseOptions.builder()
+//                .setCredentials(GoogleCredentials.getApplicationDefault())
+//                .build();
+//        FirebaseApp.initializeApp(options);
+//
+//    }
+
+    //KEY 없을 때의 예외처리 코드 필요.
+    @Bean
+    public FirebaseAuth firebaseAuth() throws IOException {
+
+        ClassPathResource key = new ClassPathResource("codingyo-firebase-adminsdk.json");
+
+        FirebaseOptions options = FirebaseOptions.builder()
+//        FirebaseOptions options = new FirebaseOptions.Builder()
+                .setCredentials(GoogleCredentials.fromStream(key.getInputStream()))
+                .build();
+        FirebaseApp.initializeApp(options);
+        return FirebaseAuth.getInstance(FirebaseApp.getInstance());
+    }
+
+}
+
+
